Changed inc file name and added new makefile
